Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,10 +2,10 @@
 
 int main() {
 
-int radius;
+float radius;
 float pi = 3.14, area;
 
-scanf("%d", &radius);
+scanf("%f", &radius);
 
 area = pi * (radius * radius);
